requirements.txt
pytest==6.2.0
Flask==1.1.2
Flask-RESTful==0.3.7
flask_cors==3.0.8
requests==2.25.0
nltk==3.5
pymongo==3.11.2
aiohttp==3.7.3
pyspellchecker==0.5.5
python-Levenshtein==0.12.0
dateparser==1.0.0
